+++ transmart-oauth/build.gradle
@@ -49,8 +49,10 @@ dependencies {
     testCompile "org.grails.plugins:geb"
     testRuntime "org.seleniumhq.selenium:selenium-htmlunit-driver:2.47.1"
     testRuntime "net.sourceforge.htmlunit:htmlunit:2.18"
+
     compile 'org.grails.plugins:spring-security-core:3.1.1'
     compile 'org.grails.plugins:spring-security-oauth2-provider:3.0.0-RC2'
+    compile 'org.springframework.security:spring-security-ldap:3.2.3.RELEASE'
 
     compile project(':search-domain')
 }
