+++ transmart-oauth/src/main/groovy/org/transmart/oauth/AuthUserDetails.groovy
@@ -1,4 +1,4 @@
-package org.transmart.oauth2
+package org.transmart.oauth
 
 import grails.plugin.springsecurity.userdetails.GrailsUser
 import grails.util.Holders
