+++ transmart-server/build.gradle
@@ -58,7 +58,6 @@ dependencies {
 
     compile 'org.grails.plugins:spring-security-core:3.1.1'
     runtime 'org.springframework.security:spring-security-config:3.2.3.RELEASE'
-    compile 'org.springframework.security:spring-security-ldap:3.2.3.RELEASE'
     compile 'org.springframework.security:spring-security-web:3.2.3.RELEASE'
     compile 'org.grails.plugins:spring-security-oauth2:1.1.0'
     compile 'org.grails.plugins:spring-security-oauth2-google:1.1.0'
