+++ transmart-server/src/main/groovy/security/BadCredentialsEventListener.groovy
@@ -4,7 +4,7 @@ import org.springframework.context.ApplicationListener
 import org.springframework.security.authentication.event.AbstractAuthenticationFailureEvent
 import org.springframework.security.authentication.event.AuthenticationFailureBadCredentialsEvent
 import org.springframework.security.authentication.event.AuthenticationFailureServiceExceptionEvent
-import org.transmart.oauth2.BruteForceLoginLockService
+import org.transmart.oauth.authentication.BruteForceLoginLockService
 
 class BadCredentialsEventListener implements ApplicationListener<AbstractAuthenticationFailureEvent> {
