+++ transmart-server/src/test/groovy/org/transmart/server/AuthControllerSpec.groovy
@@ -1,0 +1,22 @@
+package org.transmart.server
+
+import grails.test.mixin.TestFor
+import spock.lang.Specification
+
+/**
+ * See the API for {@link grails.test.mixin.web.ControllerUnitTestMixin} for usage instructions
+ */
+@TestFor(AuthController)
+class AuthControllerSpec extends Specification {
+
+    def setup() {
+    }
+
+    def cleanup() {
+    }
+
+    void "test something"() {
+        expect:"fix me"
+            true == false
+    }
+}
