--- transmartApp/grails-app/domain/org/transmart/oauth2/AccessToken.groovy
@@ -1,38 +1,0 @@
-package org.transmart.oauth2
-
-class AccessToken {
-
-    String authenticationKey
-    byte[] authentication
-
-    String username
-    String clientId
-
-    String value
-    String tokenType
-
-    Date expiration
-    Map<String, Object> additionalInformation
-
-    static hasOne = [refreshToken: String]
-    static hasMany = [scope: String]
-
-    static constraints = {
-        username nullable: true
-        clientId nullable: false, blank: false
-        value nullable: false, blank: false, unique: true
-        tokenType nullable: false, blank: false
-        expiration nullable: false
-        scope nullable: false
-        refreshToken nullable: true
-        authenticationKey nullable: false, blank: false, unique: true
-        authentication nullable: false, minSize: 1, maxSize: 1024 * 4
-        additionalInformation nullable: true
-    }
-
-    static mapping = {
-        datasource 'oauth2'
-        version false
-        scope lazy: false
-    }
-}
