--- transmartApp/grails-app/domain/org/transmart/oauth2/RefreshToken.groovy
@@ -1,19 +1,0 @@
-package org.transmart.oauth2
-
-class RefreshToken {
-
-    String value
-    Date expiration
-    byte[] authentication
-
-    static constraints = {
-        value nullable: false, blank: false, unique: true
-        expiration nullable: true
-        authentication nullable: false, minSize: 1, maxSize: 1024 * 4
-    }
-
-    static mapping = {
-        datasource 'oauth2'
-        version false
-    }
-}
